group members

+++ src/app/header/header.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, EventEmitter, OnInit, Output } from '@angular/core';
 import { CartserviceService } from '../service/cartservice.service';
 import { Router } from '@angular/router';
+import { ApiserviceService } from '../service/apiservice.service';
 
 @Component({
   selector: 'app-header',
@@ -11,13 +12,21 @@ export class HeaderComponent implements OnInit{
   searchTerm: any;
   totalItem :any;
   name: any;
-constructor(private cartservice:CartserviceService, public router:Router){}
+  chartservice: any;
+  
+  @Output() searchChanged = new EventEmitter<string>();
+
+constructor(private cartservice:CartserviceService, public router:Router, private apiservice:ApiserviceService){}
   ngOnInit() {
     let user = JSON.parse(String(localStorage.getItem("userdetails")));
 this.name=user.Name;
     this.cartservice.getproducts().subscribe(res=>{
       this.totalItem = res.lenght;
     })
+
+
+
+  
    
   }
 
@@ -26,9 +35,19 @@ search() {
 
   }
   logout(){
+  let exit = confirm("Do you want to Exit");
+  if(exit==true){
     localStorage.clear();
     this.router.navigateByUrl("/login") 
+
+  }
+
    }
 
+   onSearch(input: any) {
+    console.log(input.target.value)
+    this.searchChanged.emit(input.target.value);
+  }
 }
+
